feat: implement GetWeekSummary use case and connect weekly activity flow

+++ goalsTracker/domain/src/main/java/com/tomiappdevelopment/goalstracker/domain/modules/WeekDataSum.kt
@@ -2,5 +2,5 @@ package com.tomiappdevelopment.goalstracker.domain.modules
 
 data class WeekDataSum(
     val quantity:Int,
-    val distance:Double
+    val distance:Float
 )
